fix : 답변 삭제시 permission 확인하는 책임을 Answer로 변경(피드백 반영)

+++ src/main/java/qna/domain/Answer.java
@@ -12,6 +12,7 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.Lob;
 import javax.persistence.ManyToOne;
+import qna.CannotDeleteException;
 import qna.NotFoundException;
 import qna.UnAuthorizedException;
 
@@ -91,9 +92,16 @@ public class Answer extends BaseTimeEntity {
         return deleted;
     }
 
-    public DeleteHistory delete() {
+    public DeleteHistory delete(User loginUser) throws CannotDeleteException {
+        validateDeletePermission(loginUser);
         this.deleted = true;
         return new DeleteHistory(ContentType.ANSWER, id, writer, LocalDateTime.now());
+    }
+
+    private void validateDeletePermission(User loginUser) throws CannotDeleteException {
+        if (!isOwner(loginUser)) {
+            throw new CannotDeleteException("다른 사람이 쓴 답변이 있어 삭제할 수 없습니다.");
+        }
     }
 
     @Override
